Safari-609.1.11.1/LayoutTests/fast/selectors/read-only-read-write-designmode-basics-expected.txt

Test the basic features of ":read-only", ":read-write" on elements with the design mode enabled. The definition is that ":read-write" is matches for "elements that are editing hosts or editable and are neither input elements nor textarea elements"

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testQuerySelector() is true
PASS testStyling() is true
PASS successfullyParsed is true

TEST COMPLETE
